Read language from regkey for TSF registration
Microsoft Word has a annoying bug in which Pinyin toned vowels are
displayed in the SimSun font, while the surrounding consonants remain in
the Latin font that the user has selected.  The workaround is to
register PinyinTones under Japanese, instead of Chinese.  However,
Japanese doesn't work perfectly in all applications, either.  For
example, WordPad displays toned vowels as double-width characters!

PinyinTones now reads the HKLM\SOFTWARE\PinyinTones\TSFLanguage (DWORD)
and registers itself as a text service for this language.  The next step
will be to write an installer that presents the choices in an
understandable way and writes the regkey.