Update build and installation framework
Converted project to Visual Studio 2012, which has deprecated the Visual
Studio setup project.  Created new installers using the WiX 3.6 toolset,
sharing the same UpgradeCode between 32-bit and 64-bit.  WiX allows for
greater control over upgrades and prerequisites, preventing problems
such as installing 32-bit PinyinTones on 64-bit Windows.

Also restored the makefile for the Windows DDK build environment.  The
DDK builds against the system version of the Visual C++ runtime, and is
the approach recommended by the Microsoft TSF team.  The resulting DLL
is targeted at Vista and above, and will run as a drop-in system
component, without the need for a Visual C++ redistributable to be
installed.

This fixes a very annoying bug with Visual Studio 2010 setup projects.
Only the VC++ 2010 RTM redist package could be set as a prerequisite,
and the installer would fail if the user had the VC++ 2010 SP1 redist
package installed.