Handle only specific characters in a composition
The idea is to be in a composition as little as possible.  Only keys in
the Pinyin alphabet will start a composition, and only specific keys are
handled within a composition.  A tone number will terminate the
composition, as will any non-Pinyin key.

Previously, any key that produced a Unicode character would start a new
composition or be handled by PinyinTones in an existing composition.
This had many side effects.  The Tab key could no longer be used to tab
through a dialog, and the Delete key would leave the program in an
indeterminate state.  Numerals between 1 and 4 could not be typed.

It is much better to handle only specific keys.  Keys that are not
handled by PinyinTones, such as control keys, can be typed directly into
the application outside of a composition.  It is also possible to type
any number, since numerals are only eaten within a composition.