Clean up composition on termination
The composition must be cleaned up whenever it is terminated, including
when PinyinTones did not initiate the termination.  For example, the
composition will be terminated when a different window gets the focus.

Without clean-up, the dotted-underline display attribute will remain in
place.  After switching back to the original application, it will look
like the composition is still active, even though it has already been
terminated.

Even Microsoft IMEs have trouble with this.  The Microsoft Pinyin IME
does not clear the display attribute.  The Microsoft Japanese IME clears
the display attribute at the next keystroke, which of course starts a
new composition instead of continuing the original composition.

WordPad clears the display attribute itself.  Word does not.  Notepad is
not TSF-aware, yet it gets the best behavior of all -- the composition
actually remains active through the focus change.