Release COM pointers safely in cleanup blocks
In functions that release all COM pointers in a cleanup block, the
pointers should be initialized when declared.  Otherwise, an early exit
might cause an uninitialized pointer to be accessed and released.  Also,
early exits must be done with a goto instead of a return, so that the
cleanup block runs.

The Microsoft sample code does not need to be changed.  Some functions
release COM pointers within if blocks.  Other functions only goto the
cleanup block if a reference to the COM object has been successfully
obtained, and return early if it has not.