Allow the 'v' character to be entered
The 'v' character can now be entered by typing the 'v' key twice.

The 'v' key is ordinarily used to enter the Pinyin 'ü' character.
However, Pinyin is often mixed with text that is in another language,
such as in a vocabulary list.

It is inconvenient to have to switch to a different input method
whenever a 'v' is needed, especially if lots of input methods are
installed.  Enabling the 'v' character allows short blocks of non-Pinyin
text to be typed without switching out of PinyinTones.