Reduce maximum length of composition
Since compositions are now automatically terminated in most situations,
the maximum length of a composition can be reduced.  If the composition
reaches the maximum length, the start anchor is shifted so that part of
the text is automatically accepted.

The remaining composition only needs to be long enough to hold one
Pinyin syllable, although we leave it longer than that.